pyproject.toml
[build-system]
requires = [
    "setuptools>=45",
    "wheel",
    "scmver"
]
build-backend = "setuptools.build_meta"

[project]
name = "merlinquantum"
dynamic = ["version", "optional-dependencies"]
description = "Quantum neural network models using photonic circuits - Preview"
readme= "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "MerLin Team"},
    {name = "MerLin", email = "MerLin@quandela.com"}
]

classifiers=[
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Physics",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "torch>=2.0.0",
    "perceval-quandela>=0.13.1",
    "pandas"
]

[project.urls]
homepage = "https://merlinquantum.ai"
documentation = "https://merlinquantum.ai"
source = "https://github.com/merlinquantum/merlin"
issues = "https://github.com/merlinquantum/merlin/issues"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["merlin"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

